requirements.txt
flask
flask-smorest
python-dotenv
openai
PyJWT
Flask-SQLAlchemy
requests
rembg